add two functions

+++ src/test/java/org/onism/FactTest.java
@@ -11,9 +11,10 @@ public class FactTest {
     @Test
     public void testFact() {
         Fact demo = new Fact();
-        int[] array = {1, 5, 7, 4, 3,2, 6};
+        float[] array = {1,5, 7, 4, 3,2, 6};
+//        demo.QuickSort(array, 0, array.length - 1);
         demo.QuickSort(array, 0, array.length - 1);
-        for  (int i: array) {
+        for  (float i: array) {
             System.out.print(i + " ");
         }
     }
